backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
